third_party/WebKit/LayoutTests/fast/forms/select/popup-menu-touch-operations-expected.txt

PASS picker._selectElement.value is "1"
PASS menuElement.value is "bar"
PASS picker._selectElement.value is "0"
PASS picker._selectElement.value is "2"
PASS window.internals.pagePopupWindow is null
PASS menuElement.value is "baz"
PASS successfullyParsed is true

TEST COMPLETE
